Might be working right now

+++ frontend/src/components/CategoryTable.svelte
@@ -1,9 +1,16 @@
 <script lang="ts">
-    import { categoryStore } from "../lib/store";
-    import { onDestroy } from "svelte";
-    import type { backend } from "../../wailsjs/go/models";
-    import { DeleteCategoryById } from "../../wailsjs/go/backend/App";
-    import TrashIcon from "./TrashIcon.svelte";
+    import { categoryStore, loadProducts } from "$lib/store";
+    import { DeleteCategoryById } from "$wails/go/backend/App";
+    import type { backend } from "$wails/go/models";
+    import { onDestroy, onMount } from "svelte";
+    import TrashIcon from "$components/icons/TrashIcon.svelte";
+    import {
+        modalStore,
+        type ModalComponent,
+        type ModalSettings,
+    } from "@skeletonlabs/skeleton";
+    import UpdateCategoryModal from "./modals/UpdateCategoryModal.svelte";
+    import EditIcon from "./icons/EditIcon.svelte";
 
     let categoryList: backend.Category[] = [];
     let filteredCategoryList: backend.Category[] = [];
@@ -17,16 +24,42 @@
     });
 
     async function DeleteCategory(id: number) {
-        await DeleteCategoryById(id);
-        const index = categoryList.findIndex((product) => product.id === id);
-        if (index !== -1) {
-            categoryList.splice(index, 1);
-            filterCategories();
-        }
+        const confirm: ModalSettings = {
+            type: "confirm",
+            title: "Eliminar categoría",
+            body: "Estás seguro que deseas eliminar la categoría?",
+            buttonTextCancel: "Cancelar",
+            buttonTextConfirm: "Confirmar",
+            response: async (confirmed: boolean) => {
+                if (confirmed) {
+                    await DeleteCategoryById(id);
+                    const index = categoryList.findIndex(
+                        (category) => category.id === id
+                    );
+                    if (index !== -1) {
+                        categoryList.splice(index, 1);
+                        filterCategories();
+                    }
+                    await loadProducts();
+                }
+            },
+        };
+        modalStore.trigger(confirm);
     }
-    onDestroy(() => {
-        unsubscribe();
-    });
+
+    function EditCategory(category: backend.Category) {
+        const c: ModalComponent = {
+            ref: UpdateCategoryModal,
+            props: { category },
+        };
+        const d: ModalSettings = {
+            buttonTextCancel: "X",
+            type: "component",
+            component: c,
+            title: "Edit Category",
+        };
+        modalStore.trigger(d);
+    }
 
     function filterCategories() {
         filteredCategoryList = categoryList.filter((category) => {
@@ -46,6 +79,14 @@
             );
         });
     }
+
+    onMount(() => {
+        loadProducts();
+    });
+
+    onDestroy(() => {
+        unsubscribe();
+    });
 </script>
 
 <div class="input-group input-group-divider grid-cols-[auto_1fr_auto]">
@@ -77,12 +118,18 @@
                 {#each filteredCategoryList as category}
                     <tr>
                         <td>{category.id}</td>
-                        <td>{category.name}</td>
+                        <td class="capitalize">{category.name}</td>
                         <td>{new Date(category.updated_at).toLocaleString()}</td
                         >
                         <td>{category.profit_percent}%</td>
-                        <td
-                            ><button
+                        <td>
+                            <button
+                                class=" bg-initial hover:text-success-400"
+                                on:click={() => EditCategory(category)}
+                                ><EditIcon /></button
+                            >
+
+                            <button
                                 class=" bg-initial hover:text-error-400"
                                 on:click={() => DeleteCategory(category.id)}
                                 ><TrashIcon /></button
