+++ frontend/src/components/ProductTable.svelte
@@ -1,18 +1,23 @@
 <script lang="ts">
-    import { loadProducts, productsStore } from "../lib/store";
     import { onDestroy, onMount } from "svelte";
-    import type { backend } from "../../wailsjs/go/models";
-    import { DeleteProductById } from "../../wailsjs/go/backend/App";
-    import TrashIcon from "./TrashIcon.svelte";
-    import EditIcon from "./EditIcon.svelte";
+    import {
+        modalStore,
+        type ModalComponent,
+        type ModalSettings,
+    } from "@skeletonlabs/skeleton";
+    import { loadProducts, productsStore } from "$lib/store";
+    import type { backend } from "$wails/go/models";
+    import { DeleteProductById } from "$wails/go/backend/App";
+    import EditIcon from "./icons/EditIcon.svelte";
+    import TrashIcon from "./icons/TrashIcon.svelte";
+    import UpdateProductModal from "./modals/UpdateProductModal.svelte";
 
-    onMount(() => loadProducts());
     function roundPrice(num: number) {
         const roundedNum = Math.round(num * 2) / 2;
         return roundedNum;
     }
-    let productsList: backend.ProductWithCategory[] = [];
-    let filteredProductsList: backend.ProductWithCategory[] = [];
+    let productsList: backend.ProductWithCategoryAndProvider[] = [];
+    let filteredProductsList: backend.ProductWithCategoryAndProvider[] = [];
     let searchQuery: string = "";
     const unsubscribe = productsStore.subscribe((products) => {
         productsList = products;
@@ -22,13 +27,41 @@
     });
 
     async function DeleteProduct(id: number) {
-        await DeleteProductById(id);
-        const index = productsList.findIndex((product) => product.id === id);
-        if (index !== -1) {
-            productsList.splice(index, 1);
-            filterProducts();
-        }
-        await loadProducts();
+        const confirm: ModalSettings = {
+            type: "confirm",
+            title: "Eliminar producto",
+            body: "Estás seguro que deseas eliminar el producto?",
+            buttonTextCancel: "Cancelar",
+            buttonTextConfirm: "Confirmar",
+            response: async (confirmed: boolean) => {
+                if (confirmed) {
+                    await DeleteProductById(id);
+                    const index = productsList.findIndex(
+                        (product) => product.id === id
+                    );
+                    if (index !== -1) {
+                        productsList.splice(index, 1);
+                        filterProducts();
+                    }
+                    await loadProducts();
+                }
+            },
+        };
+        modalStore.trigger(confirm);
+    }
+
+    function EditProduct(product: backend.ProductWithCategoryAndProvider) {
+        const c: ModalComponent = {
+            ref: UpdateProductModal,
+            props: { product },
+        };
+        const d: ModalSettings = {
+            buttonTextCancel: "X",
+            type: "component",
+            component: c,
+            title: "Edit Product",
+        };
+        modalStore.trigger(d);
     }
 
     function filterProducts() {
@@ -41,7 +74,7 @@
                 product.provider
                     .toLowerCase()
                     .includes(searchQuery.toLowerCase()) ||
-                product.cost_price
+                product.unit_cost_price
                     .toString()
                     .toLowerCase()
                     .includes(searchQuery.toLowerCase()) ||
@@ -49,13 +82,19 @@
                     .toLocaleString()
                     .toLowerCase()
                     .includes(searchQuery.toLowerCase()) ||
-                product.sell_price
+                product.unit_sell_price
                     .toString()
+                    .toLowerCase()
+                    .includes(searchQuery.toLowerCase()) ||
+                product.category_name
                     .toLowerCase()
                     .includes(searchQuery.toLowerCase())
             );
         });
     }
+    onMount(() => {
+        loadProducts();
+    });
     onDestroy(() => {
         unsubscribe();
     });
@@ -82,8 +121,8 @@
                 <th>Nombre</th>
                 <th>Actualizado</th>
                 <th>Proveedor</th>
-                <th>Precio Costo</th>
-                <th>Precio Venta</th>
+                <th>Precio Costo Unitario</th>
+                <th>Precio Venta Unitario</th>
                 <th>Categoría</th>
                 <th />
             </tr>
@@ -95,9 +134,9 @@
                     <td class="capitalize">{product.name.toLowerCase()}</td>
                     <td>{new Date(product.updated_at).toLocaleString()}</td>
                     <td class="capitalize">{product.provider.toLowerCase()}</td>
-                    <td>{product.cost_price}</td>
+                    <td>{product.unit_cost_price}</td>
                     <td
-                        >{`${roundPrice(product.sell_price)}${" "}(${
+                        >{`${roundPrice(product.unit_sell_price)}${" "}(${
                             product.category_profit_percent
                         }%)`}
                     </td>
@@ -107,7 +146,7 @@
                     <td
                         ><button
                             class=" bg-initial hover:text-success-400"
-                            on:click={() => DeleteProduct(product.id)}
+                            on:click={() => EditProduct(product)}
                             ><EditIcon /></button
                         >
